Create lsl
I wanted ls -l to show permissions in octal but I couldn't find a way to do that. The Linux utility stat will show permissions in octal but it doesn't output all the other information in an easy to read way that ls -l does.
I make a ~/.bin directory that I put all my scripts in and add it to the path, so now I can use lsl anywhere on my system.